requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
discord==1.0.1
discord.py==1.5.0
emoji==0.6.0
emojis==0.6.0
flake8==3.8.3
idna==2.10
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.0
mccabe==0.6.1
multidict==4.7.6
mysql-connector==2.2.9
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
ply==3.11
psutil==5.7.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.3
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
skills==0.3.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
youtube-dl==2020.7.28
